refactor(#1526): extract lifecycle-policy coordinator (seam 6)
Extract the lifecycle-policy pipeline from orchestrator.ts into a new
LifecyclePolicyCoordinator (orchestration/lifecycle-policy-coordinator.ts).

Methods moved:
- actionOutcomePriorDelta (fully extracted — only internal caller moved)
- buildLifecycleActionPriors (fully extracted — only internal caller moved)
- runLifecyclePolicyPass (body → thin delegation)
- runFactArchival (body → thin delegation)
- runSummarization (body → thin delegation)
- runTopicExtraction (body → thin delegation)

Behavior-preserving: the orchestrator keeps thin private wrappers so all
call sites (runConsolidation, runLifecyclePolicyNow) work unchanged.

LOC: orchestrator.ts 20312 → 19957 (−355).
Entity-hardening 246/246 green.

Chokepoints used: capability gates (resolveMemoryLifecycleCapabilities,
resolveCreationMemoryCapabilities, resolveIndexingCapabilities).
