+++ scripts/ratchet-baseline.json
@@ -4,7 +4,7 @@
   "oversizeThresholdLoc": 3000,
   "metrics": {
     "watchlistLoc": {
-      "packages/remnic-core/src/orchestrator.ts": 20312,
+      "packages/remnic-core/src/orchestrator.ts": 19957,
       "packages/remnic-core/src/cli.ts": 9391,
       "packages/remnic-core/src/access-service.ts": 9013,
       "packages/remnic-core/src/storage.ts": 7747,
